requirements.txt
django
Pillow
django-tinymce